Refacto select component
