+++ x-pack/plugins/fleet/public/applications/fleet/components/fleet_server_instructions/components/index.ts
@@ -4,4 +4,4 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-export * from './fleet_server_host_combobox';
+export * from './fleet_server_hosts_form';
