+++ x-pack/plugins/fleet/public/applications/fleet/components/fleet_server_instructions/steps/get_started.tsx
@@ -5,13 +5,10 @@
  * 2.0.
  */
 
-import React, { useMemo } from 'react';
+import React from 'react';
 
-import type { EuiTheme } from '@kbn/kibana-react-plugin/common';
 import {
-  EuiIcon,
   EuiStepProps,
-  EuiSuperSelect,
   EuiButton,
   EuiCallOut,
   EuiCode,
@@ -26,13 +23,13 @@ import {
 } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
 import { FormattedMessage } from '@kbn/i18n-react';
-import { useTheme } from 'styled-components';
 
 import { MultiRowInput } from '../../../sections/settings/components/multi_row_input';
 
 import { useLink } from '../../../hooks';
 
 import type { QuickStartCreateForm } from '../hooks';
+import { FleetServerHostSelect } from '../components';
 
 export function getGettingStartedStep(props: QuickStartCreateForm): EuiStepProps {
   return {
@@ -54,38 +51,6 @@ const GettingStartedStepContent: React.FunctionComponent<QuickStartCreateForm> =
   submit,
 }) => {
   const { getHref } = useLink();
-  const theme = useTheme() as EuiTheme;
-
-  const fleetServerHostsOptions = useMemo(
-    () => [
-      ...fleetServerHosts.map((fleetServerHost) => {
-        return {
-          inputDisplay: `${fleetServerHost.name} (${fleetServerHost.host_urls[0]})`,
-          value: fleetServerHost.id,
-        };
-      }),
-      {
-        icon: <EuiIcon type="plus" size="m" color="primary" />,
-        inputDisplay: (
-          <FormattedMessage
-            id="xpack.fleet.fleetServerSetup.addFleetServerHostBtn"
-            defaultMessage="Add new Fleet Server Hosts"
-          />
-        ),
-        dropdownDisplay: (
-          <EuiText size="relative" color={theme.eui.euiColorPrimary}>
-            <FormattedMessage
-              id="xpack.fleet.fleetServerSetup.addFleetServerHostBtn"
-              defaultMessage="Add new Fleet Server Hosts"
-            />
-          </EuiText>
-        ),
-        'data-test-subj': 'fleetServerSetup.addNewHostBtn',
-        value: '@@##ADD_FLEET_SERVER_HOST##@@',
-      },
-    ],
-    [fleetServerHosts, theme.eui.euiColorPrimary]
-  );
 
   if (status === 'success') {
     return (
@@ -132,28 +97,11 @@ const GettingStartedStepContent: React.FunctionComponent<QuickStartCreateForm> =
 
       <EuiSpacer size="m" />
       {selectedFleetServerHost ? (
-        <>
-          <EuiSuperSelect
-            fullWidth
-            data-test-subj="fleetServerSetup.fleetServerHostsSelect"
-            prepend={
-              <EuiText size="relative" color={''}>
-                <FormattedMessage
-                  id="xpack.fleet.fleetServerSetup.fleetServerHostsLabel"
-                  defaultMessage="Fleet Server Hosts"
-                />
-              </EuiText>
-            }
-            onChange={(fleetServerHostId) =>
-              setFleetServerHost(
-                fleetServerHosts.find((fleetServerHost) => fleetServerHost.id === fleetServerHostId)
-              )
-            }
-            valueOfSelected={selectedFleetServerHost?.id}
-            options={fleetServerHostsOptions}
-          />
-          <EuiSpacer size="m" />
-        </>
+        <FleetServerHostSelect
+          setFleetServerHost={setFleetServerHost}
+          selectedFleetServerHost={selectedFleetServerHost}
+          fleetServerHosts={fleetServerHosts}
+        />
       ) : null}
 
       <EuiForm onSubmit={submit}>
